Add files via upload
This repository contains the solution to Assignment 2 for Software Engineering. It includes certificates for completing LinkedIn Learning courses on C++ Design Patterns and the implementation of an Ant Farm Simulation in C++. The simulation uses design patterns such as Singleton, Factory, and Decorator, providing a command-line interface for user interaction. All necessary files and resources are included for review and execution.